Move Metric annotation under MetricEvent interface. Cleanup code.
