+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/internal/navigation/RouteRetrievalEvent.kt
@@ -2,7 +2,9 @@ package com.mapbox.services.android.navigation.v5.internal.navigation
 
 import android.annotation.SuppressLint
 import android.os.Parcelable
+import com.google.gson.Gson
 import com.mapbox.navigation.base.metrics.DirectionsMetrics
+import com.mapbox.navigation.base.metrics.MetricEvent
 import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.NavigationPerformanceEvent
 
 @SuppressLint("ParcelCreator")
@@ -11,7 +13,7 @@ internal data class RouteRetrievalEvent(
     private val routeUuid: String,
     @Transient private val sessionId: String,
     @Transient override var metadata: NavigationPerformanceMetadata
-) : NavigationPerformanceEvent(sessionId, DirectionsMetrics.ROUTE_RETRIEVAL, metadata), Parcelable {
+) : NavigationPerformanceEvent(sessionId, DirectionsMetrics.ROUTE_RETRIEVAL, metadata), MetricEvent, Parcelable {
 
     companion object {
         private const val ELAPSED_TIME_NAME = "elapsed_time"
@@ -22,4 +24,9 @@ internal data class RouteRetrievalEvent(
         addCounter(DoubleCounter(ELAPSED_TIME_NAME, elapsedTime))
         addAttribute(Attribute(ROUTE_UUID_NAME, routeUuid))
     }
+
+    override fun toJson(gson: Gson): String = gson.toJson(this)
+
+    override val metric: String
+        get() = DirectionsMetrics.ROUTE_RETRIEVAL
 }
