+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/internal/navigation/metrics/NavigationDepartEvent.kt
@@ -1,7 +1,6 @@
 package com.mapbox.services.android.navigation.v5.internal.navigation.metrics
 
 import android.annotation.SuppressLint
-import com.google.gson.Gson
 import com.mapbox.navigation.base.metrics.NavigationMetrics
 
 @SuppressLint("ParcelCreator")
@@ -10,6 +9,4 @@ internal class NavigationDepartEvent(
 ) : NavigationEvent(phoneState) {
 
     override fun getEventName(): String = NavigationMetrics.DEPART
-
-    override fun toJson(gson: Gson): String = gson.toJson(this)
 }
