+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/internal/navigation/metrics/NavigationEvent.kt
@@ -2,6 +2,7 @@ package com.mapbox.services.android.navigation.v5.internal.navigation.metrics
 
 import android.os.Build
 import android.os.Parcel
+import com.google.gson.Gson
 import com.mapbox.android.telemetry.Event
 import com.mapbox.navigation.base.metrics.MetricEvent
 import com.mapbox.services.android.navigation.BuildConfig
@@ -72,4 +73,9 @@ internal abstract class NavigationEvent(
 
     override fun writeToParcel(dest: Parcel, flags: Int) {
     }
+
+    override fun toJson(gson: Gson): String = gson.toJson(this)
+
+    override val metric: String
+        get() = getEventName()
 }
