+++ libnavigation-base/src/main/java/com/mapbox/navigation/base/metrics/MetricsReporter.kt
@@ -2,7 +2,7 @@ package com.mapbox.navigation.base.metrics
 
 interface MetricsReporter {
 
-    fun addEvent(@Metric metric: String, metricEvent: MetricEvent)
+    fun addEvent(metricEvent: MetricEvent)
 
     fun setMetricsObserver(metricsObserver: MetricsObserver)
 }
